requirements.txt
# Core ML (CPU-only torch for cloud)
torch>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
joblib>=1.3.0
tqdm>=4.65.0

# App
streamlit>=1.28.0
plotly>=5.17.0

# ECG data (only needed if retraining)
wfdb>=4.1.0
pyserial>=3.5
cloudpickle>=2.2.0
matplotlib>=3.7.0